+++ packages/ormed_cli/lib/src/commands/base/runner_command.dart
@@ -11,7 +11,7 @@ abstract class RunnerCommand extends Command<void> {
     argParser.addOption(
       'config',
       abbr: 'c',
-      help: 'Path to orm.yaml (defaults to project root).',
+      help: 'Path to ormed.yaml (defaults to project root).',
     );
     argParser.addFlag(
       'preview',
@@ -36,7 +36,7 @@ abstract class RunnerCommand extends Command<void> {
     argParser.addOption(
       'connection',
       help:
-          'Select a specific connection block defined in orm.yaml (defaults to default_connection or the only entry).',
+          'Select a specific connection block defined in ormed.yaml (defaults to default_connection or the only entry).',
     );
   }