+++ packages/ormed_mysql/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.0-dev+2
+
+- Synchronized release with ormed_cli rebranding.
+
 ## 0.1.0-dev+1
 
 - Internal version bump to align with ORMed release.